refactor: route-contract parity and rounds/jsonExtract consolidation
- writersRoom/sprites/musicVideo/games list endpoints gain the same
  opt-in pagination branch universeBuilder and creativeDirector already
  use; bare requests keep returning plain arrays
- rounds and brainSongbook partial updates move from PUT to PATCH to
  match every sibling route; both client wrappers and verb-asserting
  tests updated in the same change
- the 15 rounds sanitizers + limit constants extract to
  server/lib/roundsValidation.js (barrel + README registered); rounds.js
  drops 1004→636 lines with its public exports unchanged
- universeBuilderRefine's hand-rolled brace walker rebuilds on
  lib/jsonExtract like mediaPromptRefiner already did
- writersRoom routes gain the missing objects-bible test block; games
  routes cover the previously untested update/delete/unbind surface
