+++ server/lib/index.js
@@ -37,6 +37,7 @@ export * as peerSyncValidation from './peerSyncValidation.js';
 export * as pipelineValidation from './pipelineValidation.js';
 export * as postValidation from './postValidation.js';
 export * as privacyValidation from './privacyValidation.js';
+export * as roundsValidation from './roundsValidation.js';
 export * as socketValidation from './socketValidation.js';
 export * as storyBuilderValidation from './storyBuilderValidation.js';
 export * as telegramValidation from './telegramValidation.js';
