+++ server/routes/musicVideo.test.js
@@ -80,6 +80,19 @@ describe('musicVideo routes', () => {
     expect(r.body).toEqual([{ id: 'mv-1', name: 'A' }]);
   });
 
+  it('GET / returns a bounded envelope when pagination is requested', async () => {
+    svc.listProjects.mockResolvedValueOnce(
+      Array.from({ length: 5 }, (_, i) => ({ id: `mv-${i}`, name: `P${i}` }))
+    );
+    const r = await request(app).get('/api/music-video?limit=2&offset=1');
+    expect(r.status).toBe(200);
+    expect(r.body.items).toHaveLength(2);
+    expect(r.body.items[0].id).toBe('mv-1');
+    expect(r.body.total).toBe(5);
+    expect(r.body.limit).toBe(2);
+    expect(r.body.offset).toBe(1);
+  });
+
   it('GET /:id 404s when missing', async () => {
     svc.getProject.mockResolvedValue(null);
     const r = await request(app).get('/api/music-video/mv-x');
